requirements.txt
# Core UI
streamlit

# LangChain
langchain
langchain-core
langchain-community
langchain-text-splitters

# LLM providers
langchain-groq
langchain-openai
langchain-google-genai

# Embeddings
langchain-huggingface
sentence-transformers

# Agent orchestration and memory
langgraph

# Vector store
faiss-cpu

# Document loaders
pypdf
docx2txt

# Web search
requests

# Utilities
python-dotenv
langchain-openrouter